move check bypass to disable module decorator
