+++ app/consumer/ante/disabled_modules_ante.go
@@ -5,6 +5,7 @@ import (
 	"strings"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
+	"github.com/cosmos/cosmos-sdk/x/authz"
 )
 
 type DisabledModulesDecorator struct {
@@ -20,12 +21,52 @@ func (dmd DisabledModulesDecorator) AnteHandle(ctx sdk.Context, tx sdk.Tx, simul
 	for _, msg := range tx.GetMsgs() {
 		msgTypeURL := sdk.MsgTypeURL(msg)
 
-		for _, prefix := range dmd.prefixes {
-			if strings.HasPrefix(msgTypeURL, prefix) {
-				return ctx, fmt.Errorf("tx contains message types from unsupported modules at height %d", currHeight)
+		if hasDisabledModuleMsgs(msg, dmd.prefixes...) {
+			return ctx, fmt.Errorf("tx contains message types from unsupported modules at height %d", currHeight)
+		}
+
+		// Check if there is an atempt to bypass disabled module msg
+		// with authz MsgExec
+		if msgTypeURL == "/cosmos.authz.v1beta1.MsgExec" {
+			wrappedMsgs := fetchMsgExecWrapperMsgs(msg)
+			for _, wrappedMsg := range wrappedMsgs {
+				if hasDisabledModuleMsgs(wrappedMsg, dmd.prefixes...) {
+					return ctx, fmt.Errorf("tx contains message types from unsupported modules at height %d", currHeight)
+				}
 			}
 		}
 	}
 
 	return next(ctx, tx, simulate)
+}
+
+func fetchMsgExecWrapperMsgs(msg sdk.Msg) []sdk.Msg {
+	var wrappedMsgs = []sdk.Msg{}
+
+	msgExec, ok := msg.(*authz.MsgExec)
+
+	if !ok {
+		return []sdk.Msg{}
+	}
+
+	sdkMsgs, err := msgExec.GetMessages()
+	if err != nil {
+		return []sdk.Msg{}
+	}
+
+	wrappedMsgs = append(wrappedMsgs, sdkMsgs...)
+
+	return wrappedMsgs
+}
+
+func hasDisabledModuleMsgs(msg sdk.Msg, prefixes ...string) bool {
+	msgType := sdk.MsgTypeURL(msg)
+
+	for _, prefix := range prefixes {
+		if strings.HasPrefix(msgType, prefix) {
+			return true
+		}
+	}
+
+	return false
 }
