+++ app/consumer/ante/msg_filter_ante.go
@@ -5,7 +5,6 @@ import (
 	"strings"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
-	"github.com/cosmos/cosmos-sdk/x/authz"
 )
 
 type (
@@ -18,14 +17,12 @@ type (
 	// filtering based on certain criteria.
 	MsgFilterDecorator struct {
 		ConsumerKeeper ConsumerKeeper
-		prefixes       []string
 	}
 )
 
-func NewMsgFilterDecorator(k ConsumerKeeper, disabledModules ...string) MsgFilterDecorator {
+func NewMsgFilterDecorator(k ConsumerKeeper) MsgFilterDecorator {
 	return MsgFilterDecorator{
 		ConsumerKeeper: k,
-		prefixes:       disabledModules,
 	}
 }
 
@@ -38,13 +35,6 @@ func (mfd MsgFilterDecorator) AnteHandle(ctx sdk.Context, tx sdk.Tx, simulate bo
 		if !hasValidMsgsPreCCV(tx.GetMsgs()) {
 			return ctx, fmt.Errorf("tx contains unsupported message types at height %d", currHeight)
 		}
-	}
-
-	// Check if there is an atempt to bypass disabled module msg
-	// with authz MsgExec
-	wrappedMsgs := fetchMsgExecWrapperMsgs(tx.GetMsgs())
-	if hasDisabledModuleMsgs(wrappedMsgs, mfd.prefixes...) {
-		return ctx, fmt.Errorf("tx contains message types from unsupported modules at height %d", currHeight)
 	}
 
 	return next(ctx, tx, simulate)
@@ -63,44 +53,4 @@ func hasValidMsgsPreCCV(msgs []sdk.Msg) bool {
 	}
 
 	return true
-}
-
-func fetchMsgExecWrapperMsgs(msgs []sdk.Msg) []sdk.Msg {
-	var wrappedMsgs = []sdk.Msg{}
-	for _, msg := range msgs {
-		msgType := sdk.MsgTypeURL(msg)
-
-		if msgType == "/cosmos.authz.v1beta1.MsgExec" {
-			msgExec, ok := msg.(*authz.MsgExec)
-
-			// TODO: We should decide to continue checking the other msgs or break the loop
-			if !ok {
-				continue
-			}
-
-			sdkMsgs, err := msgExec.GetMessages()
-			if err != nil {
-				continue
-			}
-
-			wrappedMsgs = append(wrappedMsgs, sdkMsgs...)
-
-		}
-	}
-
-	return wrappedMsgs
-}
-
-func hasDisabledModuleMsgs(msgs []sdk.Msg, prefixes ...string) bool {
-	for _, msg := range msgs {
-		msgType := sdk.MsgTypeURL(msg)
-
-		for _, prefix := range prefixes {
-			if strings.HasPrefix(msgType, prefix) {
-				return true
-			}
-		}
-	}
-
-	return false
 }
